Found issue with construction of production trees
As the resources needed increase, so do the number of possible ways you can make the resource. For example, when constructing computers there are over 9 trillion possible combinations. This presents an issue in that you cannot possibly calculate every production path for computers, but you also can't make each required resource the same way throughout the tree as that approach would not allow for the elegant handling of byproducts in production